update linter rules, fix warnings

+++ node/blockchain/fullblocktests/generate.go
@@ -537,9 +537,9 @@ func (g *testGenerator) blocksByMMMRoot(mmrRoot chainhash.Hash) *wire.MsgBlock {
 // - A coinbase that pays the required subsidy to an OP_TRUE script
 // - When a spendable output is provided:
 //   - A transaction that spends from the provided output the following outputs:
-//     - One that pays the inputs amount minus 1 atom to an OP_TRUE script
-//     - One that contains an OP_RETURN output with a random uint64 in order to
-//       ensure the transaction has a unique hash
+//   - One that pays the inputs amount minus 1 atom to an OP_TRUE script
+//   - One that contains an OP_RETURN output with a random uint64 in order to
+//     ensure the transaction has a unique hash
 //
 // Additionally, if one or more munge functions are specified, they will be
 // invoked with the block prior to solving it.  This provides callers with the
@@ -644,7 +644,7 @@ func (g *testGenerator) nextBlock(blockName string, spend *spendableOut, mungers
 	var leafsInCurMMR []*mmr.TreeNode
 
 	for i := int32(0); i <= numLeafsInCurMMR; i++ {
-		leafsInCurMMR = append(leafsInCurMMR, g.mmr.Block(int32(i)))
+		leafsInCurMMR = append(leafsInCurMMR, g.mmr.Block(i))
 	}
 
 	g.blockNameToMMRTreeLeafs[blockName] = leafsInCurMMR
